requirements.txt
fastmcp>=2.5.1
gradio[mcp]>=5.31.0
openai>=1.82.0
textblob>=0.19.0
httpx
python-dotenv
fastapi
uvicorn
